Use FC and add some tests
