+++ ts/components/screens/BaseScreenComponent/index.tsx
@@ -1,0 +1,234 @@
+import { fromPredicate } from "fp-ts/lib/Option";
+import { BugReporting } from "instabug-reactnative";
+import { Millisecond } from "italia-ts-commons/lib/units";
+import { Container } from "native-base";
+import { connectStyle } from "native-base-shoutem-theme";
+import mapPropsToStyleNames from "native-base/src/utils/mapPropsToStyleNames";
+import React, { useState, useEffect, ComponentProps } from "react";
+import { useSelector } from "react-redux";
+import { ColorValue, ModalBaseProps, Platform } from "react-native";
+
+import { GlobalState } from "../../../store/reducers/types";
+import { getCurrentRouteName } from "../../../utils/navigation";
+import { TranslationKeys } from "../../../../locales/locales";
+import {
+  DefaultReportAttachmentTypeConfiguration,
+  defaultAttachmentTypeConfiguration
+} from "../../../boot/configureInstabug";
+import I18n from "../../../i18n";
+import customVariables from "../../../theme/variables";
+import { setStatusBarColorAndBackground } from "../../../utils/statusBar";
+import ContextualHelp, { RequestAssistancePayload } from "../../ContextualHelp";
+import { SearchType } from "../../search/SearchButton";
+import Markdown from "../../ui/Markdown";
+import { noAnalyticsRoutes } from "../../../utils/analytics";
+import { mixpanelTrack } from "../../../mixpanel";
+import { AccessibilityEvents, BaseHeader } from "../BaseHeader";
+
+import { handleOnContextualHelpDismissed, handleOnLinkClicked } from "./utils";
+
+export interface ContextualHelpProps {
+  title: string;
+  body: () => React.ReactNode;
+}
+
+export interface ContextualHelpPropsMarkdown {
+  title: TranslationKeys;
+  body: TranslationKeys;
+}
+
+interface OwnProps {
+  onAccessibilityNavigationHeaderFocus?: () => void;
+  accessibilityEvents?: AccessibilityEvents;
+  accessibilityLabel?: string;
+  contextualHelp?: ContextualHelpProps;
+  contextualHelpMarkdown?: ContextualHelpPropsMarkdown;
+  headerBody?: React.ReactNode;
+  headerBackgroundColor?: ColorValue;
+  appLogo?: boolean;
+  searchType?: SearchType;
+  reportAttachmentTypes?: DefaultReportAttachmentTypeConfiguration;
+
+  // As of now, the following prop is propagated through 4 levels
+  // to finally display a checkbox in SendSupportRequestOptions
+  shouldAskForScreenshotWithInitialValue?: boolean;
+}
+
+export type Props = OwnProps &
+  ComponentProps<typeof BaseHeader> &
+  Pick<ComponentProps<typeof ContextualHelp>, "faqCategories">;
+
+const maybeDark = fromPredicate(
+  (isDark: boolean | undefined = undefined) => isDark === true
+);
+
+const ANDROID_OPEN_REPORT_DELAY = 50 as Millisecond;
+
+const contextualHelpModalAnimation = Platform.select<
+  ModalBaseProps["animationType"]
+>({
+  ios: "slide",
+  default: "none"
+});
+
+const BaseScreenComponentFC: React.FC<Props> = ({
+  accessibilityEvents,
+  accessibilityLabel,
+  appLogo,
+  children,
+  contextualHelp,
+  contextualHelpMarkdown,
+  customGoBack,
+  customRightIcon,
+  dark,
+  faqCategories,
+  goBack,
+  headerBackgroundColor,
+  headerBody,
+  headerTitle,
+  isSearchAvailable,
+  onAccessibilityNavigationHeaderFocus,
+  primary,
+  reportAttachmentTypes,
+  shouldAskForScreenshotWithInitialValue,
+  showInstabugChat,
+  titleColor
+}) => {
+  const currentScreenName = useSelector((store: GlobalState) =>
+    getCurrentRouteName(store.nav)
+  );
+
+  const [isHelpVisible, setIsHelpVisible] = useState(false);
+  // if the content is markdown we listen for load end event, otherwise the content is
+  // assumed always loaded
+  const [markdownContentLoaded, setMarkdownContentLoaded] = useState<boolean>(
+    !contextualHelpMarkdown
+  );
+
+  // used to trigger the side-effect base on timeout to take the screenshot
+  const [requestAssistanceData, setRequestAssistanceData] = useState<{
+    payload: RequestAssistancePayload;
+    attachmentConfig: DefaultReportAttachmentTypeConfiguration;
+  } | null>(null);
+
+  useEffect(() => {
+    if (requestAssistanceData) {
+      // since in Android we have no way to handle Modal onDismiss event https://reactnative.dev/docs/modal#ondismiss
+      // we force handling here. The timeout is due to wait until the modal is completely hidden
+      // otherwise in the Instabug screeshot we will see the contextual help content instead the screen below
+      // TODO: To complete the porting to 0.63.x, both iOS and Android will use the timeout. https://www.pivotaltracker.com/story/show/174195300
+      setTimeout(() => {
+        handleOnContextualHelpDismissed(
+          requestAssistanceData.payload,
+          requestAssistanceData.attachmentConfig
+        );
+        setRequestAssistanceData(null);
+      }, ANDROID_OPEN_REPORT_DELAY);
+    }
+  }, [requestAssistanceData]);
+
+  const showHelp = () => {
+    void mixpanelTrack("OPEN_CONTEXTUAL_HELP", {
+      SCREEN_NAME: noAnalyticsRoutes.has(currentScreenName)
+        ? "_"
+        : currentScreenName
+    });
+
+    maybeDark(dark).map(_ =>
+      setStatusBarColorAndBackground("dark-content", customVariables.colorWhite)
+    );
+    setIsHelpVisible(true);
+    setMarkdownContentLoaded(!contextualHelpMarkdown);
+  };
+
+  const hideHelp = () => {
+    maybeDark(dark).map(_ =>
+      setStatusBarColorAndBackground(
+        "light-content",
+        customVariables.brandDarkGray
+      )
+    );
+    setIsHelpVisible(false);
+  };
+
+  const handleOnRequestAssistance = (payload: RequestAssistancePayload) => {
+    // if reportAttachmentTypes is undefined use the default attachment config
+    const attachmentConfig: DefaultReportAttachmentTypeConfiguration = {
+      ...(reportAttachmentTypes ?? defaultAttachmentTypeConfiguration),
+      screenshot: payload.shouldSendScreenshot ?? true
+    };
+    if (payload.supportType === BugReporting.reportType.bug) {
+      setIsHelpVisible(false);
+      setRequestAssistanceData({ payload, attachmentConfig });
+    } else {
+      // don't close modal if the report isn't a bug (bug brings a screenshot)
+      handleOnContextualHelpDismissed(payload, attachmentConfig);
+    }
+  };
+
+  const contextualHelpConfig = contextualHelp
+    ? { body: contextualHelp.body, title: contextualHelp.title }
+    : contextualHelpMarkdown
+    ? {
+        body: () => (
+          <Markdown
+            onLinkClicked={handleOnLinkClicked(hideHelp)}
+            onLoadEnd={() => {
+              setMarkdownContentLoaded(true);
+            }}
+          >
+            {I18n.t(contextualHelpMarkdown.body)}
+          </Markdown>
+        ),
+        title: I18n.t(contextualHelpMarkdown.title)
+      }
+    : undefined;
+
+  return (
+    <Container>
+      <BaseHeader
+        onAccessibilityNavigationHeaderFocus={
+          onAccessibilityNavigationHeaderFocus
+        }
+        accessibilityEvents={accessibilityEvents}
+        accessibilityLabel={accessibilityLabel}
+        showInstabugChat={showInstabugChat}
+        primary={primary}
+        dark={dark}
+        goBack={goBack}
+        headerTitle={headerTitle}
+        backgroundColor={headerBackgroundColor}
+        onShowHelp={contextualHelpConfig ? showHelp : undefined}
+        isSearchAvailable={isSearchAvailable}
+        body={headerBody}
+        appLogo={appLogo}
+        customRightIcon={customRightIcon}
+        customGoBack={customGoBack}
+        titleColor={titleColor}
+      />
+      {children}
+      {contextualHelpConfig && (
+        <ContextualHelp
+          shouldAskForScreenshotWithInitialValue={
+            shouldAskForScreenshotWithInitialValue
+          }
+          title={contextualHelpConfig.title}
+          onLinkClicked={handleOnLinkClicked(hideHelp)}
+          body={contextualHelpConfig.body}
+          isVisible={isHelpVisible}
+          modalAnimation={contextualHelpModalAnimation}
+          onRequestAssistance={handleOnRequestAssistance}
+          close={hideHelp}
+          contentLoaded={markdownContentLoaded}
+          faqCategories={faqCategories}
+        />
+      )}
+    </Container>
+  );
+};
+
+export default connectStyle(
+  "UIComponent.BaseScreenComponent",
+  {},
+  mapPropsToStyleNames
+)(BaseScreenComponentFC);
